requirements.txt
openai>=1.0.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
pypdf==4.3.1
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
pandas==2.2.2
scikit-learn==1.5.1
python-dotenv==1.0.1